pyproject.toml
[project]
name = "quarterback"
version = "1.0.0"
description = "Strategic task prioritization and agent orchestration for multi-project operators"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Bobby Goldsmith" }]
keywords = ["mcp", "task-management", "prioritization", "agent-orchestration", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0.0"]
import = ["beautifulsoup4>=4.12.0", "requests>=2.31.0"]
pro = []
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.1.0"]

[project.scripts]
quarterback = "quarterback.cli:main"
quarterback-server = "quarterback.server:run"

[project.urls]
Homepage = "https://github.com/bobbyrgoldsmith/quarterback"
Repository = "https://github.com/bobbyrgoldsmith/quarterback"
Issues = "https://github.com/bobbyrgoldsmith/quarterback/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quarterback"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
